add backend.py and move functions about add task into backend.py
